+++ pdf/lib/widgets.dart
@@ -23,6 +23,7 @@ import 'package:meta/meta.dart';
 import 'package:pdf/pdf.dart';
 import 'package:vector_math/vector_math_64.dart';
 
+part 'widgets/annotations.dart';
 part 'widgets/basic.dart';
 part 'widgets/clip.dart';
 part 'widgets/container.dart';